% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopQTL-class.R
\docType{class}
\name{loopQTL-class}
\alias{loopQTL-class}
\alias{samples}
\alias{loops}
\alias{cis}
\alias{vcfPath}
\alias{resolution}
\alias{phenotype}
\alias{results}
\alias{metadata}
\alias{samples,loopQTL-method}
\alias{loops,loopQTL-method}
\alias{cis,loopQTL-method}
\alias{vcfPath,loopQTL-method}
\alias{resolution,loopQTL-method}
\alias{phenotype,loopQTL-method}
\alias{results,loopQTL-method}
\alias{metadata,loopQTL-method}
\alias{show,loopQTL-method}
\title{Class to store loopQTL input data and analysis results}
\usage{
\S4method{samples}{loopQTL}(object)

\S4method{loops}{loopQTL}(object)

\S4method{cis}{loopQTL}(object)

\S4method{vcfPath}{loopQTL}(object)

\S4method{resolution}{loopQTL}(object)

\S4method{phenotype}{loopQTL}(object)

\S4method{results}{loopQTL}(object)

\S4method{metadata}{loopQTL}(object)

\S4method{show}{loopQTL}(object)
}
\arguments{
\item{object}{a loopQTL object}
}
\description{
loopQTL objects should normally be generated by \code{loopQTLsetup()}. See its
documentation for more information on the expected format of input data.
}
\details{
All slots can also be accessed through an identically named function, i.e.,
to access \code{results(obj)}, use \code{loopQTL::results(obj)}.
}
\section{Slots}{

\describe{
\item{\code{samples}}{List of unique sample IDs}

\item{\code{loops}}{Consensus intra-chromosomal loops as a
\link[InteractionSet:GInteractions-class]{InteractionSet::GInteractions} object}

\item{\code{cis}}{Optional numeric vector of per-sample cis-contact totals. Must
have names matching the IDs in \code{samples}, and be in the same order as
\code{samples}.}

\item{\code{vcfPath}}{Path to the tabix-indexed VCF}

\item{\code{resolution}}{Analysis bin size (bp)}

\item{\code{phenotype}}{Matrix containing normalized numeric phenotype data for each
loop (in the rows) and each sample (in the columns). Dimensions and order
has not been run.}

\item{\code{results}}{QTL result table. Empty if \code{runQTL()} has not been run.
Contains a row for each loop-SNP pair and columns \code{loopID}, \code{snpID},
\code{chrom}, \code{pos}, \code{ref}, \code{alt}, \code{anchor}, \code{beta}, \code{se}, \code{t}, \code{p}, \code{nUsed},
and \code{maf}. If \code{loopQTLannotate()} has been run, also contains \code{gene},
\code{peak} (optional), and \code{Type} columns.}

\item{\code{metadata}}{Function-specific metadata}
}}

\examples{
obj <- loopQTLExampleData()
obj
samples(obj)
loops(obj)
}
\seealso{
loopQTLsetup
}
