% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopQTLpairData.R
\name{loopQTLpairData}
\alias{loopQTLpairData}
\title{Sample-level view of one loop × SNP pair}
\usage{
loopQTLpairData(obj, loopID, snpID, normal = TRUE)
}
\arguments{
\item{obj}{A loopQTL object with \code{results(obj)} filled.}

\item{loopID}{Character vector of loopIDs to extract.}

\item{snpID}{Character vector of snpIds to extract.}

\item{normal}{Apply RINT per loop before returning phenotype values. default \code{TRUE}.}
}
\value{
data.frame with columns \code{sample}, \code{loopID}, \code{snpID}, \code{genotype}
(0 - Reference homozygou; 1 - heterozygous, 2 - Alternative homozygous), \code{phenotype}, \code{cis}.
}
\description{
This function returns the per-sample phenotype and genotype
(dosage 0/1/2) for a single SNP-loop pair. Useful for inspecting a
specific hit or exporting sample-level values for reporting.
}
\examples{
## Load the example object. In real analysis, the input
## should be the output of runQTL().
##Here we just use the example object. 
obj <- loopQTLExampleData()
##Check results so that you can choose your interested loopID and snpID
head(results(obj))
##Then you can specify the specifi SNP-loop pair data you want to check at samples level.
pair <- loopQTLpairData(
    obj,
    loopID = "L0000187_chr1_46030000_46220000",
    snpID  = "rs11584814")
head(pair)
}
