% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopQTLplotPair.R
\name{loopQTLplotPair}
\alias{loopQTLplotPair}
\title{Per-sample boxplot for a single SNP-loop pair}
\usage{
loopQTLplotPair(obj, loopID, snpID, normal = TRUE, jitter = TRUE)
}
\arguments{
\item{obj}{A loopQTL object with \code{results(obj)} filled by \code{runQTL()}.}

\item{loopID}{A single loopID to plot.}

\item{snpID}{A single snpID to plot.}

\item{normal}{Plot Rank-based Inverse Normal Transformed (RINT'd) phenotype
values. Defaults to TRUE. Set to \code{FALSE} to plot pre-RINT normalized values.}

\item{jitter}{Add a jittered point overlay (default \code{TRUE}).}
}
\value{
A \code{ggplot} object.
}
\description{
This function renders the per-sample boxplot for a chosen result SNP-loop
pair. It is returned as a ggplot object that can be saved with
\code{ggplot2::ggsave()}.
}
\examples{
## Load the example object. In real analysis, the input
## should be the output of runQTL().
##Here we just use the example object. 
obj <- loopQTLExampleData()
##Check results so that you can choose your interested loopID and snpID
head(results(obj))
##Then you can specify the specifi SNP-loop pair data you want to plot as classical QTL box plot.
##Return is a ggplot object so you can save with ggsave().
p <- loopQTLplotPair(
    obj,
    loopID = "L0000187_chr1_46030000_46220000",
    snpID  = "rs11584814")
}
